requirements.txt
anyio==3.6.2
asgiref==3.4.1
audioread==3.0.1
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
contextlib2==21.6.0
contextvars==2.4
dataclasses==0.8
decorator==4.4.2
fastapi==0.83.0
ffmpeg-python==0.2.0
flask==2.0.3
future==1.0.0
h11==0.13.0
idna==3.10
imageio==2.15.0
imageio-ffmpeg==0.5.1
immutables==0.19
importlib-metadata==4.8.3
itsdangerous==2.0.1
jinja2==3.0.3
joblib==1.1.1
librosa==0.7.0
llvmlite==0.31.0
markupsafe==2.0.1
numba==0.48.0
numpy==1.19.5
opencv-contrib-python==4.2.0.34
opencv-python==4.1.0.25
pillow==8.4.0
pip==21.2.2
proglog==0.1.10
pycparser==2.21
pydantic==1.9.2
python==3.6.13
requests==2.27.1
resampy==0.3.1
scikit-learn==0.24.2
scipy==1.5.4
setuptools==58.0.4
six==1.16.0
sniffio==1.2.0
soundfile==0.12.1
sqlite==3.45.3
starlette==0.19.1
threadpoolctl==3.1.0
torch==1.10.2
torchvision==0.11.3
tqdm==4.45.0
typing-extensions==4.1.1
urllib3==1.26.20
uvicorn==0.16.0
vc==14.40
vs2015_runtime==14.40.33807
werkzeug==2.0.3
wheel==0.37.1
wincertstore==0.2
zipp==3.6.0
